refactoring pass

+++ ElevatorSim.hpp
@@ -51,7 +51,7 @@
 #include <cmath>
 
 const float M_PI = 3.141592f;
-const double FPS = (double)1/(double)60;
+const double FPS = 1/(double)60;
 
 struct Vec3f {
 	float x, y, z;
